requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
appdirs==1.4.4
asgiref==3.3.4
attrs==21.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
distlib==0.3.2
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
filelock==3.0.12
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.3
httpx==0.18.1
idna==2.10
iniconfig==1.1.1
mysql-connector==2.2.9
numpy==1.20.3
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
Pillow==8.3.1
pluggy==0.13.1
protobuf==3.17.0
py==1.10.0
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.1
toml==0.10.2
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
typing-extensions==3.10.0.0
urllib3==1.26.4
virtualenv==20.4.7
